+++ src/table/src/lib.rs
@@ -23,9 +23,7 @@ pub mod stats;
 pub mod table;
 pub mod table_reference;
 pub mod test_util;
-pub mod thin_table;
 
 pub use crate::error::{Error, Result};
 pub use crate::stats::{ColumnStatistics, TableStatistics};
-pub use crate::table::TableRef;
-pub use crate::thin_table::ThinTable;
+pub use crate::table::{Table, TableRef};
