+++ src/table/src/table/numbers.rs
@@ -32,7 +32,7 @@ use store_api::storage::ScanRequest;
 use crate::metadata::{
     FilterPushDownType, TableId, TableInfoBuilder, TableInfoRef, TableMetaBuilder, TableType,
 };
-use crate::thin_table::ThinTable;
+use crate::table::Table;
 use crate::TableRef;
 
 const NUMBER_COLUMN: &str = "number";
@@ -50,7 +50,7 @@ impl NumbersTable {
 
     pub fn table_with_name(table_id: TableId, name: String) -> TableRef {
         let data_source = Arc::new(NumbersDataSource::new(Self::schema()));
-        let thin_table = ThinTable::new(
+        let thin_table = Table::new(
             Self::table_info(table_id, name, "test_engine".to_string()),
             FilterPushDownType::Unsupported,
             data_source,
