+++ src/table/src/test_util/empty_table.rs
@@ -21,7 +21,7 @@ use store_api::data_source::DataSource;
 use store_api::storage::ScanRequest;
 
 use crate::metadata::{FilterPushDownType, TableInfo};
-use crate::thin_table::ThinTable;
+use crate::table::Table;
 use crate::TableRef;
 
 pub struct EmptyTable;
@@ -31,7 +31,7 @@ impl EmptyTable {
         let data_source = Arc::new(EmptyDataSource {
             schema: info.meta.schema.clone(),
         });
-        let thin_table = ThinTable::new(
+        let thin_table = Table::new(
             Arc::new(info.clone()),
             FilterPushDownType::Unsupported,
             data_source,
